ntp_shirley Shirley's test as programmed at the NTP

ntp_shirley Shirley's test as programmed at the NTP

ntp_shirley(formula, data, dose_name = "dose")

Arguments

formula

An equation of the form \(Y \sim X.\) Here the variable \(Y\) is the response of interest, and \(X\) represents discrete experimental conditions. For example, if weight is the dependent variable, and you are interested in looking at the trend across sex one would have 'weight ~ sex'.

data

A data frame with column names in the formula.

dose_name

The name of the variable containing the doses in the data frame \(data\). It is expected multiple doses for each of the experimental conditions \(X\).

Value

The results of a non-parametric Shirley's isotone test for trend on each level in \(dose_name\). For more information see: doi:10.2307/2529789

The returned list contains:

X: this represents all the class objects on the right hand side of \( Y \sim X\) above.

dose: the dose groups relative to control.

testStats: Value of the Shirley test statistic.

mult_comp_signif: Test's significance as 0, 1, or 2 which is not-significant, significant at the 0.05

mult_comp_test: The type of test, i.e. "SHIRLEY"

Examples

a = ntp_shirley(weight ~ sex, data=ntp_weight_data)
summary(a)
#> Shriley's Trend Test: Monotone Change from control?  
#> ---------------------------------------------------  
#>  sex dose Significant
#>    f  125          No
#>    f  250          No
#>    f  500          No
#>    f 1000       <0.05
#>    f 2000       <0.01
#>    m  125          No
#>    m  250          No
#>    m  500          No
#>    m 1000       <0.01
#>    m 2000       <0.01